[refactor]: 파일 이름, 함수 이름 수정 - service.ts -> sessionStorage.ts - Data 객체 내부에 있던 sessionStorage 조작 메서드 모두 분리 - sessionStorage 조작 함수 이름 명시적으로 수정 (#208)
